requirements.txt
appdirs==1.4.4
argcomplete==3.6.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.1.0
python-barcode==0.15.1
python-dotenv==1.0.1
python-escpos==3.1
PyYAML==6.0.2
qrcode==8.0
six==1.17.0
Werkzeug==3.1.3
